fix timestamp and succeeded loading tasks
