+++ components/helpers/mutations.js
@@ -19,3 +19,18 @@ export const SIGNUP_MUTATION = gql`
     }
   }
 `
+
+export const CREATE_TASKS = gql`
+  mutation CreateTask($task: TaskInput) {
+    createTask(task: $task) {
+      id
+      task_name
+      task_description
+      date
+      time_start
+      time_finished
+      completed
+      user_id
+    }
+  }
+`
