+++ redux/slices/storageSlice.js
@@ -1,17 +1,18 @@
 import { createSlice, PayloadAction } from "@reduxjs/toolkit";
 
+// {
+//   title: "Test Task Title",
+//   description: "Task Description",
+//   startTime: 1670547604000,
+//   endTime: 1670548888888,
+// }
 
 const initialState = {
   username: '',
   user_id: null,
   avatar: '',
   tasks: {
-    daily: [{
-      title: "Test Task Title",
-      description: "Task Description",
-      startTime: 1670547604000,
-      endTime: 1670548888888,
-    }],
+    daily: [],
     weekly: [],
     yearly: [],
     all: [],
